Upload UNET maverick detector
An attempt at the maverick detector. Need to upload data and comment the file. Basically the UNET model except it has to take in dimensions for data that does not resemble an image. Achieved halfway decent accuracy but not close enough to be useful. The loss function might be the answer?